refactor(searchutil) ♻️ extract search utils
